+++ include/core/managers/BaseManager.hpp
@@ -1,8 +1,12 @@
 #pragma once
 #include <memory>
+#include <vector>
 #include "core/GameObject.hpp"
 
 class BaseManager{
+protected:
+    std::vector<std::weak_ptr<GameObject>> _ManagedObjects;
+
 public:
     BaseManager();
     virtual ~BaseManager();