+++ src/core/managers/EnemyManager.cpp
@@ -4,22 +4,22 @@ EnemyManager::EnemyManager() {}
 EnemyManager::~EnemyManager() {}
 
 void EnemyManager::Init() {
-    _Enemies.clear();
+    _ManagedObjects.clear();
 }
 void EnemyManager::Bind(std::shared_ptr<GameObject> enemy){
-    _Enemies.push_back(enemy);
+    _ManagedObjects.push_back(enemy);
 }
 void EnemyManager::Update(float deltaTime) {
-    for(auto& enemy : _Enemies){
+    for(auto& enemy : _ManagedObjects){
         if(auto sharedEnemy = enemy.lock()){
             if(sharedEnemy->IsActive()){
-                sharedEnemy->UpdateAutomatic(deltaTime);
+                sharedEnemy->UpdateControlled(deltaTime);
             }
         }
     }
 }
 bool EnemyManager::IsEmpty(){
-    for(auto& enemy : _Enemies){
+    for(auto& enemy : _ManagedObjects){
         if(auto sharedEnemy = enemy.lock()){
             if(sharedEnemy->IsActive()){
                 return false;